#8 Refactor express API
