+++ src/backend/package.json
@@ -2,9 +2,9 @@
   "name": "backend",
   "version": "1.0.0",
   "private": true,
-  "main": "src/app.js",
+  "main": "src/index.js",
   "scripts": {
-    "start": "node src/app"
+    "start": "node src/index"
   },
   "dependencies": {
     "axios": "^1.4.0",
@@ -13,6 +13,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
+    "express-validator": "^7.0.1",
     "firebase-admin": "^11.8.0",
     "google-auth-library": "^9.1.0",
     "lodash": "^4.17.21",
