+++ src/backend/src/routes/custom-game.routes.js
@@ -1,0 +1,36 @@
+const express = require("express");
+const { body, param } = require("express-validator");
+const {
+  isValidPlaylistId,
+  isValidTimezoneString,
+  validateInputs,
+} = require("../utils/utils");
+const CustomGameController = require("../controllers/custom-game.controllers");
+
+const router = express.Router();
+
+router.get(
+  "/:playlistId/dailySong",
+  isValidPlaylistId(param("playlistId")),
+  isValidTimezoneString("timeZone"),
+  validateInputs,
+  CustomGameController.getDailySong
+);
+
+router.get(
+  "/:playlistId/allSongs",
+  isValidPlaylistId(param("playlistId")),
+  validateInputs,
+  CustomGameController.getAllSongs
+);
+
+router.post(
+  "/:playlistId/stats",
+  isValidPlaylistId(param("playlistId")),
+  isValidTimezoneString("timeZone"),
+  body("score").isInt({ min: 0, max: 6 }),
+  validateInputs,
+  CustomGameController.postStats
+);
+
+module.exports = router;
