+++ src/backend/src/utils/firebase.utils.js
